+++ cmem_plugin_shapes/plugin_shapes.py
@@ -119,10 +119,11 @@ def str2bool(value: str) -> bool:
             param_type=BoolParameterType(),
             name="prefix_cc",
             label="Fetch namespace prefixes from prefix.cc",
-            description="Attempt to fetch the list of namespace prefixes from https://prefix.cc "
-            "instead of using the local database. If this fails, fall back on the local database. "
-            "Prefixes defined in the Corporate Memory project will override prefixes defined in "
-            "the database.",
+            description="Fetch the list of namespace prefixes from https://prefix.cc instead of "
+            "using the local prefix database. If unavailable, fall back to the local database. "
+            "Prefixes defined in the Corporate Memory project override database prefixes. Enabling "
+            "this option exposes your IP address to prefix.cc but no other data is shared. If "
+            "unsure, keep this option disabled. See https://prefix.cc/about.",
             advanced=True,
         ),
         PluginParameter(